pyproject.toml
[project]
name = "numerologprofnew"
version = "0.1.0"
description = "New numerology reports"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fpdf2>=2.8.3",
    "glom>=24.11.0",
    "loguru>=0.7.3",
    "pyright>=1.1.402",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
